Minor modifications to make the code compatible with VC++ 2012.

+++ include/PgnParserException.hpp
@@ -2,21 +2,22 @@
 #define INCLUDE_PGN_PARSER_EXCEPTION_HPP
 
 #include <stdexcept>
+#include <cstdint>
 
 namespace PgnLib
 {
   class PgnParserException : public std::runtime_error
   {
   private:
-    uint64_t line_number_;
+    int32_t line_number_;
 
   public:
     inline PgnParserException(std::string what)
       : std::runtime_error(what),
         line_number_(-1) {};
 
-    inline uint64_t line_number() const { return line_number_; };
-    inline void set_line_number(uint64_t line_number) { line_number_ = line_number; };
+    inline int32_t line_number() const { return line_number_; };
+    inline void set_line_number(int32_t line_number) { line_number_ = line_number; };
   };
 }
